requirements.txt
pywebview
faster-whisper